Thread.SetApartmentState only works on Windows
Added a condition to setting the ApartmentState.STA to only be triggerred on Windows machine. Preparation for future solution cleanup and project dvision.
Entry point to the app no longer needs to be Windows app (!) only a small library with Win32 stuff, render form etc, needs to be, and can be simply consumed via reference.