Continued WIP of ValueResult
Benchmarks and small fixes to allocations.
Need more focus on redducing closures in lambdas - focus on static delegate implementations, and methods like .Bind will require overloads that take more parameters. For now plan is to support up to 8 parameters. Maybe it could be source generated, or generated by an AI.