WIP ValueResult without value
A work in progress work on adding a ValueResult implementation that does
not hold any value on success, only an error on a failure, indented to
be used in places where now a true/false is returned to indicate
success, for example in `bool Initialize` methods.

An interface and the class implementing it was added. What is now left
are the extensions, especially bind from/to `ValueResult<TValue,
TError>`, as well as any `MapError`, and `Match` functions.

After that any additional extensions, as well as proper reformatting of
code - in particular split the code into proper classes/files as those
are getting unwieldly already.