Playing more with ValueResults
Playing with an idea to allow logging when creating a `TypedValueResult` error. For now an `ILogger` is passed to a validation method, and a new extension method, for example `LogAndReturnInvaludOperation` is used to both log a message, as well construct given `ErrorResult`. For test purposes an overload for both a simple action for logging only a static text, as well as a tuple for 2 parameters, to log 2 parameter log message, were tested. Source generated logging methods were used.

TODO a more tests, including some benchmarks how passing those `ILogger` and especially `ValueTuple`s will affect performance and allocations, especially since the error creation is for non-happy path, so shouldn't allocate if it a happy-path is reached.