Playing some more with Value Errors
Benchmarking `LogAndReturn*` Value Errors methods, and debating whether
there is an actual added value in having `TypedValueError`s..
Currently the only viable implementation uses non-ref-struct typed
errors, which leads to allocations on the heap.
A generic `TypedValueError<T>` would work, but that poses an issue with
all `.Bind()`, `.Map()` etc methods, as each subsequent could return a
different error type on error, but should pass the original one if an
error appeared higher up the call chain. Currently I see no way of
achieving that with the generic approach.

Meanwhile, with `LogAndReturn*` approach, maybe having a typed error
does not really matter, as the exception won't have proper stack trace
anyway, being thrown in an aribtrary place, at the end of the call
chain. A message is already provided, and that should be enough. So
instead, if an exception is needed to be thrown, maybe some generic
`ValueErrorException` could be thrown instead, that would have stack
trace with the end of the call chain..